Initial logic for calculating metrics laid
Now varying the acceleration should give interesting results. That
should be a vector laid out depending on the binned genome merged with
sv_calls. Will move the actual calculations to rust depending on the
scale. Then will make a shiny app to show the results as if it is
happening in real time. People need to select cells by seeing 0.5 of all
the svs in the cell.