rust enters the scene
Made the basic structs and can read in the tsv.gz file as a polars df.
Now will populate the structs with columns from the df and will lay in
the impl for the game logic with the structs. Can do basic filter,
select very easily with rust code. Let's go!